Address some issues found in lgtm
Caught a few potentially issues with overflows, as well as use of
unsafe function localtime.

+++ src/core/array.c
@@ -64,7 +64,7 @@ void janet_array_ensure(JanetArray *array, int32_t capacity, int32_t growth) {
     Janet *newData;
     Janet *old = array->data;
     if (capacity <= array->capacity) return;
-    int64_t new_capacity = capacity * growth;
+    int64_t new_capacity = ((int64_t) capacity) * growth;
     if (new_capacity > INT32_MAX) new_capacity = INT32_MAX;
     capacity = (int32_t) new_capacity;
     newData = realloc(old, capacity * sizeof(Janet));
